feat: add automatically determined source chooser along with manual selection
