+++ src/main/java/frc/robot/RobotContainer.java
@@ -30,7 +30,6 @@ import frc.robot.auto.reef.Branch.Level;
 import frc.robot.auto.reef.Reef;
 import frc.robot.auto.source.SourceChooser;
 import frc.robot.commands.autonomous.autos.DynamicAuto;
-import frc.robot.commands.autonomous.autos.DynamicAutoV2;
 import frc.robot.commands.drive.DriveCommands;
 import frc.robot.subsystems.drive.Drive;
 import frc.robot.subsystems.drive.DriveConstants;
@@ -155,8 +154,8 @@ public class RobotContainer {
     private void configureButtonBindings() {
         // Default command, normal field-relative drive
 
-        drive.setDefaultCommand(DriveCommands.joystickDrive(
-                drive, () -> -controller.getRightY(), () -> -controller.getRightX(), () -> -controller.getLeftX()));
+        // drive.setDefaultCommand(DriveCommands.joystickDrive(
+        //         drive, () -> -controller.getRightY(), () -> -controller.getRightX(), () -> -controller.getLeftX()));
 
         // Switch to X pattern when X button is pressed
         controller.x().onTrue(Commands.runOnce(drive::stopWithX, drive));
@@ -184,7 +183,7 @@ public class RobotContainer {
      * @return the command to run in autonomous
      */
     public Command getAutonomousCommand() {
-        return new DynamicAutoV2(reef, sourceChooser);
+        return new DynamicAuto(reef, sourceChooser);
     }
 
     public void resetSimulationField() {
