+++ src/main/java/frc/robot/auto/source/SourceChooser.java
@@ -4,32 +4,35 @@
 
 package frc.robot.auto.source;
 
+import com.pathplanner.lib.auto.AutoBuilder;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import frc.robot.util.PoseUtil;
+import java.util.function.Supplier;
 
 /* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
 public class SourceChooser {
-    SendableChooser<Pose2d> sourcePosChooser = new SendableChooser<>();
+    SendableChooser<Supplier<Pose2d>> sourcePosChooser = new SendableChooser<>();
     /** Creates a new SourceChooser. */
     public SourceChooser() {
-        sourcePosChooser.setDefaultOption("Left Far", SourceLocations.SOURCE_LEFT_FAR);
-        sourcePosChooser.addOption("Left Middle", SourceLocations.SOURCE_LEFT);
-        sourcePosChooser.addOption("Left Close", SourceLocations.SOURCE_LEFT_CLOSE);
-        sourcePosChooser.addOption("Right Close", SourceLocations.SOURCE_RIGHT_CLOSE);
-        sourcePosChooser.addOption("Right Middle", SourceLocations.SOURCE_RIGHT);
-        sourcePosChooser.addOption("Right Far", SourceLocations.SOURCE_RIGHT_FAR);
+        sourcePosChooser.setDefaultOption("Auto", SourceLocations.getClosestSource());
+        sourcePosChooser.addOption("Left Far", () -> SourceLocations.SOURCE_LEFT_FAR);
+        sourcePosChooser.addOption("Left Middle", () -> SourceLocations.SOURCE_LEFT);
+        sourcePosChooser.addOption("Left Close", () -> SourceLocations.SOURCE_LEFT_CLOSE);
+        sourcePosChooser.addOption("Right Close", () -> SourceLocations.SOURCE_RIGHT_CLOSE);
+        sourcePosChooser.addOption("Right Middle", () -> SourceLocations.SOURCE_RIGHT);
+        sourcePosChooser.addOption("Right Far", () -> SourceLocations.SOURCE_RIGHT_FAR);
         SmartDashboard.putData("Source Pos", sourcePosChooser);
         // Use addRequirements() here to declare subsystem dependencies.
     }
 
     public Pose2d getSourcePose() {
-        return sourcePosChooser.getSelected();
+        return sourcePosChooser.getSelected().get();
     }
 
-    public SendableChooser<Pose2d> getSourceChooser() {
+    public SendableChooser<Supplier<Pose2d>> getSourceChooser() {
         return sourcePosChooser;
     }
 
@@ -47,5 +50,13 @@ public class SourceChooser {
                 PoseUtil.offsetPose(SOURCE_RIGHT, 0.5, 0); // TODO fill in actual values
         public static final Pose2d SOURCE_RIGHT_FAR =
                 PoseUtil.offsetPose(SOURCE_RIGHT, -0.5, 0); // TODO fill in actual values
+
+        public static Supplier<Pose2d> getClosestSource() {
+
+            return () -> (PoseUtil.getDistance(AutoBuilder.getCurrentPose(), SOURCE_LEFT)
+                            < PoseUtil.getDistance(AutoBuilder.getCurrentPose(), SOURCE_RIGHT)
+                    ? SOURCE_LEFT
+                    : SOURCE_RIGHT);
+        }
     }
 }
